Update main.py
Found a potential source of bias in the way I was consuming recommendations. The issue was that we we're accounting for recommendations after market hours in the smartPricing() function alone (i.e. get the price of the next time we can trade). This lead to an opening in determineBuySell() where our determineBuySell() function might consider recommendations that took place after market hours in buying and selling, and the date passed to smartPricing() might take place during marketHours, hence a case of buying when we shouldn't.